refactor: rename to adhere to PEP 8
