+++ cognee/eval_framework/corpus_builder/task_getters/TaskGetters.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import Callable, Awaitable, List
 from cognee.api.v1.cognify.cognify import get_default_tasks
-from cognee.modules.pipelines.tasks.Task import Task
+from cognee.modules.pipelines.tasks.task import Task
 from cognee.eval_framework.corpus_builder.task_getters.get_cascade_graph_tasks import (
     get_cascade_graph_tasks,
 )
