+++ cognee/modules/pipelines/__init__.py
@@ -1,3 +1,3 @@
-from .tasks.Task import Task
+from .tasks.task import Task
 from .operations.run_tasks import run_tasks
 from .operations.run_parallel import run_tasks_parallel
